Updating GenerateSalt method to be more random
Originally the RNG was being seeded with DateTime.Now.Ticks, which isn't
guaranteed to be different when called multiple times within a very small
time span. This was causing the GenerateSalt method to sometimes generate
the same salt multiple times.

A decent work around is to seed it with the hashcode of a randomly
generated GUID.